+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/messages/MessageListViewModel.kt
@@ -291,7 +291,6 @@ public class MessageListViewModel(
      * @param message Message to delete.
      */
     @JvmOverloads
-    @Suppress("ConvertArgumentToSet")
     public fun deleteMessage(message: Message, hard: Boolean = false) {
         messageListController.deleteMessage(message, hard)
     }
@@ -302,7 +301,6 @@ public class MessageListViewModel(
      *
      * @param message Message to delete.
      */
-    @Suppress("ConvertArgumentToSet")
     public fun flagMessage(
         message: Message,
         reason: String?,
@@ -388,6 +386,30 @@ public class MessageListViewModel(
      */
     public fun removeShadowBanFromUser(userId: String) {
         messageListController.removeShadowBanFromUser(userId)
+    }
+
+    /**
+     * Flags a user identified by the provided ID.
+     *
+     * @param userId The ID of the user to flag.
+     * @param reason The reason for flagging the user.
+     * @param customData Additional key-value data submitted with the request.
+     */
+    public fun flagUser(
+        userId: String,
+        reason: String? = null,
+        customData: Map<String, String> = emptyMap(),
+    ) {
+        messageListController.flagUser(userId = userId, reason = reason, customData = customData)
+    }
+
+    /**
+     * Un-flags a user identified by the provided ID.
+     *
+     * @param userId The ID of the user to un-flag.
+     */
+    public fun unflagUser(userId: String) {
+        messageListController.unflagUser(userId)
     }
 
     /**